move useConnectionQualityIndicator hook
