+++ packages/react/src/components/participant/ConnectionQualityIndicator.tsx
@@ -1,35 +1,13 @@
 import * as React from 'react';
-import { setupConnectionQualityIndicator } from '@livekit/components-core';
-import { useEnsureParticipant } from '../../context';
-import type { Participant } from 'livekit-client';
-import { ConnectionQuality } from 'livekit-client';
 import { mergeProps } from '../../utils';
 import { getConnectionQualityIcon } from '../../assets/icons/util';
-import { useObservableState } from '../../hooks/internal/useObservableState';
-
-/** @public */
-export interface ConnectionQualityIndicatorOptions {
-  participant?: Participant;
-}
+import type { ConnectionQualityIndicatorOptions } from '../../hooks';
+import { useConnectionQualityIndicator } from '../../hooks';
 
 /** @public */
 export interface ConnectionQualityIndicatorProps
   extends React.HTMLAttributes<HTMLDivElement>,
     ConnectionQualityIndicatorOptions {}
-
-/** @public */
-export function useConnectionQualityIndicator(options: ConnectionQualityIndicatorOptions = {}) {
-  const p = useEnsureParticipant(options.participant);
-
-  const { className, connectionQualityObserver } = React.useMemo(
-    () => setupConnectionQualityIndicator(p),
-    [p],
-  );
-
-  const quality = useObservableState(connectionQualityObserver, ConnectionQuality.Unknown);
-
-  return { className, quality };
-}
 
 /**
  * The ConnectionQualityIndicator shows the individual connection quality of a participant.
