+++ packages/react/src/hooks/useConnectionQualityIndicator.ts
@@ -1,0 +1,25 @@
+import { setupConnectionQualityIndicator } from '@livekit/components-core';
+import type { Participant } from 'livekit-client';
+import { ConnectionQuality } from 'livekit-client';
+import React from 'react';
+import { useEnsureParticipant } from '../context';
+import { useObservableState } from './internal';
+
+/** @public */
+export interface ConnectionQualityIndicatorOptions {
+  participant?: Participant;
+}
+
+/** @public */
+export function useConnectionQualityIndicator(options: ConnectionQualityIndicatorOptions = {}) {
+  const p = useEnsureParticipant(options.participant);
+
+  const { className, connectionQualityObserver } = React.useMemo(
+    () => setupConnectionQualityIndicator(p),
+    [p],
+  );
+
+  const quality = useObservableState(connectionQualityObserver, ConnectionQuality.Unknown);
+
+  return { className, quality };
+}
